+++ osu.Game/Online/API/Requests/Responses/Comment.cs
@@ -6,8 +6,6 @@ using osu.Game.Users;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net;
-using System.Text.RegularExpressions;
 
 namespace osu.Game.Online.API.Requests.Responses
 {
@@ -70,11 +68,9 @@ namespace osu.Game.Online.API.Requests.Responses
 
         public bool IsDeleted => DeletedAt.HasValue;
 
-        public bool HasMessage => !string.IsNullOrEmpty(MessageHtml);
+        public bool HasMessage => !string.IsNullOrEmpty(Message);
 
         public bool IsVoted { get; set; }
-
-        public string GetMessage => HasMessage ? WebUtility.HtmlDecode(Regex.Replace(MessageHtml, @"<(.|\n)*?>", string.Empty)) : string.Empty;
 
         public int DeletedChildrenCount => ChildComments.Count(c => c.IsDeleted);
     }