+++ osu.Game/Overlays/Comments/CommentsContainer.cs
@@ -12,7 +12,6 @@ using osu.Game.Online.API.Requests.Responses;
 using System.Threading;
 using System.Linq;
 using osu.Framework.Extensions.IEnumerableExtensions;
-using osu.Game.Users;
 
 namespace osu.Game.Overlays.Comments
 {
@@ -23,8 +22,6 @@ namespace osu.Game.Overlays.Comments
 
         public readonly Bindable<CommentsSortCriteria> Sort = new Bindable<CommentsSortCriteria>();
         public readonly BindableBool ShowDeleted = new BindableBool();
-
-        protected readonly Bindable<User> User = new Bindable<User>();
 
         [Resolved]
         private IAPIProvider api { get; set; }
@@ -71,6 +68,7 @@ namespace osu.Game.Overlays.Comments
                         },
                         new Container
                         {
+                            Name = @"Footer",
                             RelativeSizeAxes = Axes.X,
                             AutoSizeAxes = Axes.Y,
                             Children = new Drawable[]
@@ -111,8 +109,6 @@ namespace osu.Game.Overlays.Comments
                     }
                 }
             });
-
-            User.BindTo(api.LocalUser);
         }
 
         protected override void LoadComplete()
@@ -152,13 +148,14 @@ namespace osu.Game.Overlays.Comments
             loadCancellation?.Cancel();
             request = new GetCommentsRequest(type, id.Value, Sort.Value, currentPage++);
             request.Success += onSuccess;
-            api.PerformAsync(request);
+            api.Queue(request);
         }
 
         private void clearComments()
         {
             currentPage = 1;
             deletedCommentsCounter.Count.Value = 0;
+            moreButton.Show();
             moreButton.IsLoading = true;
             content.Clear();
         }
@@ -167,25 +164,10 @@ namespace osu.Game.Overlays.Comments
         {
             loadCancellation = new CancellationTokenSource();
 
-            var page = new FillFlowContainer
+            LoadComponentAsync(new CommentsPage(response)
             {
-                RelativeSizeAxes = Axes.X,
-                AutoSizeAxes = Axes.Y,
-                Direction = FillDirection.Vertical,
-            };
-
-            foreach (var c in response.Comments)
-            {
-                if (c.IsTopLevel)
-                {
-                    page.Add(new DrawableComment(c)
-                    {
-                        ShowDeleted = { BindTarget = ShowDeleted }
-                    });
-                }
-            }
-
-            LoadComponentAsync(page, loaded =>
+                ShowDeleted = { BindTarget = ShowDeleted }
+            }, loaded =>
             {
                 content.Add(loaded);
 
@@ -199,10 +181,12 @@ namespace osu.Game.Overlays.Comments
                     moreButton.Current.Value = response.TopLevelCount - loadedTopLevelComments;
                     moreButton.IsLoading = false;
                 }
+                else
+                {
+                    moreButton.Hide();
+                }
 
                 commentCounter.Current.Value = response.Total;
-
-                moreButton.FadeTo(response.HasMore ? 1 : 0);
             }, loadCancellation.Token);
         }